Move to thread_local instance for runtime
The Future trait makes it impossible to pass the runtime instance to task code by argument.

Previous way was to store an Rc<Runtime> into task closures.
This was annoying, and required storing additional pointers everywhere.

Use a thread_local instance like everyone else.
This allows removing all references to the runtime in various structs.
The runtime can be referenced from inner async blocks without difficulty.

Only one instance can live at a time for each thread, but this is not
limiting as only one instance can be used per thread.
This is due to the fact that block_on is blocking.