Ensure tasks are not scheduled multiple times
Some combinator may cause multiple wakers to be registered for a single task.
Example : join!(), or any combinator that polls multiple futures on each poll().

In pathological cases, this could generate many duplicate reference to
task in the ready_queue, using up a lot of memory for nothing.

Another risk is a completed task being scheduled.

Thus a flag was added in the task frame to detect if it is scheduled,
and prevent multiple scheduling.