Start implementation of instrusive chain.
Based on Givy's intrusive list design, but with more basic operations (no general splicing, only insert / unlink).
This is possible to express with a safe interface due to the Pin guarantees.
However this requires unsafe everywhere.

A big difference with Givy is that a pinned RawLink cannot be built in place.
Thus it cannot be initialized as a singleton with prev == next == &self.
It starts with an "unlinked" state (~ nullptr), and is linked on demand when insert is called.
As insert requires a now Pinned RawLink, this should be safe.

Added impl of Raw links that should be correct.
Added skeleton of high level Chain (list head) and Link (elements).