Split borrows to executor and reactor
During normal use, the global executor and reactor may be borrowed at the same time.
This happens mostly when the reactor wakes tasks.
The previous design used an exclusive borrow to the runtime which panicked.
As the two borrows are done in an uncoordinated way, there is no way to share the single runtime borrow.

Changed the Runtime structure design to provide dynamic borrow of executor and reactor separately.
This fixes the problem.

Also changed the way "enabling" works internally, by just using a bool flag.
This prevents having to borrow through two refcell as would require the previous design.
Cleanup of runtime after block_on is done manually.

Made the executor inline and the reactor boxed.
Boxing the reactor is required to provide pinning guarantees.
Doing it only for the reactor prevents adding indirection for the executor (most commonly used structure).